"improved" greedagent
I spent the whole afternoon doing this.

GreedAgent is now able to think and infer what a card might be based on cards that have been played (counting them). If there are more than 1 fuse tokens left and it has a 50% chance of having a playable card, it will risk playing that card.